+++ app/src/components/details/Measurements.tsx
@@ -1,8 +1,8 @@
 import React from "react";
-import { IMeasurement } from "../serverApi/IMeasurement";
+import { IMeasurement } from "../../serverApi/IMeasurement";
 import { DataGrid } from "./dataGrid/DataGrid";
 
-export const MeasurementsList: React.FC<{ measurements: IMeasurement[] }> = ({
+export const Measurements: React.FC<{ measurements: IMeasurement[] }> = ({
   measurements,
 }) => {
   return <DataGrid measurements={measurements} />;
