+++ app/src/components/details/Summary.tsx
@@ -1,5 +1,5 @@
 import React from "react";
 
-export const MetricSummary: React.FC = () => {
+export const Summary: React.FC = () => {
   return <div>Summary</div>;
 };
