+++ app/src/components/details/chart/IVisualizationProps.tsx
@@ -1,4 +1,4 @@
-import { IMeasurement } from "../../serverApi/IMeasurement";
+import { IMeasurement } from "../../../serverApi/IMeasurement";
 
 export interface IVisualizationProps {
   measurements: IMeasurement[];
