+++ app/src/components/details/chart/LazyChartJs.tsx
