+++ app/src/components/details/chart/Visualization.tsx
