+++ app/src/components/details/dataGrid/IDataGridProps.tsx
@@ -1,4 +1,4 @@
-import { IMeasurement } from "../../serverApi/IMeasurement";
+import { IMeasurement } from "../../../serverApi/IMeasurement";
 
 export interface IDataGridProps {
   measurements: IMeasurement[];
